requirements.txt
crewai
crewai[tools]
langchain-openai
python-dotenv
tavily-python